fix(embed): response.ok checks + extract shared init-api Sentry helper
Addresses Copilot + prfectionist review on #612:

1. The three single-fetch entry files (doc-verification, enhanced-
   document-verification, product-selection) only captured to Sentry
   when fetch itself rejected or JSON parsing threw. HTTP 4xx/5xx with
   a parseable body silently returned undefined and propagated as a
   generic downstream error — exactly the bug class this PR set out
   to surface. Add `response.ok` checks that throw an Error with the
   status attached so the existing catch tags it with `httpStatus`.

2. Extract the duplicated init-API capture block (basic-kyc,
   biometric-kyc, ekyc) into `init-api-sentry.js`. The three files
   used to carry an identical ~20-line block constructing failedRequests
   + productsConfigStatus + servicesStatus tags inline; they now call
   `buildInitApiFailure(...)` and `captureInitApiFailure(e, failure)`
   so future tag changes happen in one place.
