+++ packages/embed/src/js/biometric-kyc.js
@@ -20,6 +20,10 @@ import {
   idInfoToIdSelection,
   applyIdInfoPrefill,
 } from './id-info-utils.js';
+import {
+  buildInitApiFailure,
+  captureInitApiFailure,
+} from './init-api-sentry.js';
 
 (function biometricKyc() {
   'use strict';
@@ -139,32 +143,13 @@ import {
           generalConstraints: generalConstraints.hosted_web.biometric_kyc,
         };
       }
-      const failedRequests = [];
-      if (!productsConfigResponse.ok) failedRequests.push('products_config');
-      if (!servicesResponse.ok) failedRequests.push('services');
-      initApiFailure = {
-        failedRequests,
-        productsConfigStatus: productsConfigResponse.status,
-        servicesStatus: servicesResponse.status,
-      };
+      initApiFailure = buildInitApiFailure(
+        productsConfigResponse,
+        servicesResponse,
+      );
       throw new Error('Failed to get supported ID types');
     } catch (e) {
-      Sentry.captureException(e, {
-        tags: {
-          area: 'init_api',
-          failedRequest: initApiFailure
-            ? initApiFailure.failedRequests.join(',')
-            : 'rejection',
-          ...(initApiFailure
-            ? {
-                productsConfigStatus: String(
-                  initApiFailure.productsConfigStatus,
-                ),
-                servicesStatus: String(initApiFailure.servicesStatus),
-              }
-            : {}),
-        },
-      });
+      captureInitApiFailure(e, initApiFailure);
       throw new Error('Failed to get supported ID types', { cause: e });
     }
   }
