+++ packages/embed/src/js/enhanced-document-verification.js
@@ -83,12 +83,25 @@ function applyPageTranslations() {
 
     try {
       const response = await fetch(url.toString());
+      // `fetch` only rejects on network errors — 4xx/5xx still resolve, so
+      // an HTTP failure here would parse JSON of an error body and silently
+      // return undefined without a Sentry event. Explicitly throw so the
+      // catch tags the request and the status before re-throwing.
+      if (!response.ok) {
+        const err = new Error('Failed to get supported ID types');
+        err.httpStatus = response.status;
+        throw err;
+      }
       const json = await response.json();
 
       return json.hosted_web.enhanced_document_verification;
     } catch (e) {
       Sentry.captureException(e, {
-        tags: { area: 'init_api', failedRequest: 'services' },
+        tags: {
+          area: 'init_api',
+          failedRequest: 'services',
+          ...(e.httpStatus ? { httpStatus: String(e.httpStatus) } : {}),
+        },
       });
       throw new Error('Failed to get supported ID types', { cause: e });
     }
